Support builtin types; re-organize code
