+++ pydantic/validate_call_decorator.py
@@ -4,6 +4,7 @@ from __future__ import annotations as _annotations
 
 import functools
 import inspect
+from types import BuiltinFunctionType, BuiltinMethodType, FunctionType, LambdaType, MethodType
 from typing import TYPE_CHECKING, Any, Callable, TypeVar, overload
 
 from ._internal import _typing_extra, _validate_call
@@ -14,6 +15,43 @@ if TYPE_CHECKING:
     from .config import ConfigDict
 
     AnyCallableT = TypeVar('AnyCallableT', bound=Callable[..., Any])
+
+# This should be aligned with `GenerateSchema.match_types`
+_validate_call_supported_types: tuple[type[Any], ...] = (
+    LambdaType,
+    FunctionType,
+    MethodType,
+    BuiltinFunctionType,
+    BuiltinMethodType,
+    functools.partial,
+)
+
+
+def _check_function_type(function: object):
+    if any(isinstance(function, t) for t in _validate_call_supported_types):
+        try:
+            assert callable(function)  # for type checking
+            inspect.signature(function)
+        except ValueError:
+            # partial doesn't have a qualname, so we just not include it in the error message
+            maybe_qualname = f'`{function.__qualname__}` ' if hasattr(function, '__qualname__') else ''
+            raise TypeError(f"Input function {maybe_qualname}doesn't have a valid signature")
+        return
+
+    if isinstance(function, (classmethod, staticmethod)):
+        name = type(function).__name__
+        raise TypeError(f'The `@{name}` decorator should be applied after `@validate_call` (put `@{name}` on top)')
+
+    if inspect.isclass(function):
+        raise TypeError(
+            '`validate_call` should be applied to functions, not classes (put `@validate_call` on top of `__init__` or `__new__` instead)'
+        )
+    if callable(function):
+        raise TypeError(
+            '`validate_call` should be applied to functions, not instances or other callables. Use `validate_call` explicitly on `__call__` instead.'
+        )
+
+    raise TypeError('`validate_call` should be applied to one of the following: function, method, partial, or lambda')
 
 
 @overload
@@ -50,21 +88,10 @@ def validate_call(
     local_ns = _typing_extra.parent_frame_namespace()
 
     def validate(function: AnyCallableT) -> AnyCallableT:
+        _check_function_type(function)
+
         validate_call_wrapper = _validate_call.wrap_validate_call(function, config, validate_return, local_ns)
-
-        if inspect.iscoroutinefunction(function):
-
-            @functools.wraps(function)
-            async def wrapper_function(*args, **kwargs):  # type: ignore
-                return await validate_call_wrapper(*args, **kwargs)
-        else:
-
-            @functools.wraps(function)
-            def wrapper_function(*args, **kwargs):
-                return validate_call_wrapper(*args, **kwargs)
-
-        wrapper_function.raw_function = function  # type: ignore
-        return wrapper_function  # type: ignore
+        return validate_call_wrapper  # type: ignore
 
     if func:
         return validate(func)
